move add and remove helper to mod.rs
